Do packet parsing with the new code
