+++ neqo-transport/src/packet.rs
@@ -5,521 +5,852 @@
 // except according to those terms.
 
 // Encoding and decoding packets off the wire.
-
-// A lot of methods and types contain the word Packet
-#![allow(clippy::module_name_repetitions)]
-
-use neqo_common::{hex, matches, qtrace, Decoder};
-
-use std::convert::TryFrom;
-
-use crate::cid::{ConnectionId, ConnectionIdDecoder};
-use crate::crypto::CryptoDxState;
+use crate::cid::{ConnectionId, ConnectionIdDecoder, ConnectionIdRef};
+use crate::crypto::{CryptoDxState, CryptoStates};
 use crate::tracking::PNSpace;
-use crate::{Error, Res};
+use crate::{Error, Res, Version, QUIC_VERSION};
+
+use neqo_common::{hex, qdebug, qerror, qtrace, Decoder, Encoder};
+use neqo_crypto::{aead::Aead, hkdf, random, TLS_AES_128_GCM_SHA256, TLS_VERSION_1_3};
+
+use std::cell::RefCell;
+use std::convert::TryFrom;
+use std::fmt;
+use std::iter::ExactSizeIterator;
+use std::ops::{Deref, DerefMut, Range};
+use std::time::Instant;
 
 const PACKET_TYPE_INITIAL: u8 = 0x0;
 const PACKET_TYPE_0RTT: u8 = 0x01;
 const PACKET_TYPE_HANDSHAKE: u8 = 0x2;
 const PACKET_TYPE_RETRY: u8 = 0x03;
 
+const PACKET_BIT_LONG: u8 = 0x80;
+const PACKET_BIT_SHORT: u8 = 0x00;
+const PACKET_BIT_KEY_PHASE: u8 = 0x04;
+const PACKET_BIT_FIXED_QUIC: u8 = 0x40;
+
 const SAMPLE_SIZE: usize = 16;
 
-pub trait HeaderProtectionMask {
-    fn compute_mask(&self, sample: &[u8]) -> Res<Vec<u8>>;
-    #[must_use]
-    fn next_pn(&self) -> PacketNumber;
-}
+pub type PacketNumber = u64;
 
-pub trait Unprotector: HeaderProtectionMask {
-    fn decrypt(&mut self, pn: PacketNumber, hdr: &[u8], body: &[u8]) -> Res<Vec<u8>>;
-}
-
-#[derive(Debug, PartialEq)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum PacketType {
-    Short(bool), // Includes whether the key phase is set
-    ZeroRTT,
+    VersionNegotiation,
+    Initial,
     Handshake,
-    VN(Vec<Version>), // List of versions
-    Initial(Vec<u8>), // Token
-    Retry { odcid: ConnectionId, token: Vec<u8> },
+    ZeroRtt,
+    Retry,
+    Short,
+    OtherVersion,
 }
 
 impl PacketType {
     #[must_use]
-    pub fn space(&self) -> PNSpace {
+    fn code(self) -> u8 {
         match self {
-            Self::Short(_) | Self::ZeroRTT => PNSpace::ApplicationData,
-            Self::Handshake => PNSpace::Handshake,
-            Self::Initial(_) => PNSpace::Initial,
-            _ => panic!("don't ask for the space when there isn't one"),
+            Self::Initial => PACKET_TYPE_INITIAL,
+            Self::ZeroRtt => PACKET_TYPE_0RTT,
+            Self::Handshake => PACKET_TYPE_HANDSHAKE,
+            Self::Retry => PACKET_TYPE_RETRY,
+            _ => panic!("shouldn't be here"),
+        }
+    }
+}
+
+/// The AEAD used for Retry is fixed, so use this.
+fn make_retry_aead() -> Aead {
+    #[cfg(debug_assertions)]
+    ::neqo_crypto::assert_initialized();
+
+    let secret = hkdf::import_key(
+        TLS_VERSION_1_3,
+        TLS_AES_128_GCM_SHA256,
+        &[
+            0x65, 0x6e, 0x61, 0xe3, 0x36, 0xae, 0x94, 0x17, 0xf7, 0xf0, 0xed, 0xd8, 0xd7, 0x8d,
+            0x46, 0x1e, 0x2a, 0xa7, 0x08, 0x4a, 0xba, 0x7a, 0x14, 0xc1, 0xe9, 0xf7, 0x26, 0xd5,
+            0x57, 0x09, 0x16, 0x9a,
+        ],
+    )
+    .unwrap();
+    Aead::new(TLS_VERSION_1_3, TLS_AES_128_GCM_SHA256, &secret, "quic ").unwrap()
+}
+thread_local!(static RETRY_AEAD: RefCell<Aead> = RefCell::new(make_retry_aead()));
+fn retry_expansion() -> usize {
+    if let Ok(ex) = RETRY_AEAD.try_with(|aead| aead.borrow().expansion()) {
+        ex
+    } else {
+        panic!("Unable to access Retry AEAD")
+    }
+}
+
+struct PacketBuilderoffsets {
+    /// The bits of the first octet that need masking.
+    first_byte_mask: u8,
+    /// The offset of the length field.
+    len: usize,
+    /// The location of the packet number field.
+    pn: Range<usize>,
+}
+
+/// A packet builder that can be used to produce short packets and long packets.
+/// This does not produce Retry or Version Negotiation.
+pub struct PacketBuilder {
+    encoder: Encoder,
+    pn: PacketNumber,
+    header: Range<usize>,
+    offsets: PacketBuilderoffsets,
+}
+
+impl PacketBuilder {
+    /// Start building a long header packet.
+    pub fn short(mut encoder: Encoder, key_phase: bool, dcid: &ConnectionId) -> Self {
+        let header_start = encoder.len();
+        // TODO(mt) randomize the spin bit
+        encoder.encode_byte(PACKET_BIT_SHORT | PACKET_BIT_FIXED_QUIC | (u8::from(key_phase) << 2));
+        encoder.encode(&dcid);
+        Self {
+            encoder,
+            pn: u64::max_value(),
+            header: header_start..header_start,
+            offsets: PacketBuilderoffsets {
+                first_byte_mask: 0x1f,
+                pn: 0..0,
+                len: 0,
+            },
         }
     }
 
+    /// Start building a long header packet.
+    /// For an Initial packet you will need to call initial_token(),
+    /// even if the token is empty.
+    pub fn long(
+        mut encoder: Encoder,
+        pt: PacketType,
+        dcid: &ConnectionId,
+        scid: &ConnectionId,
+    ) -> Self {
+        let header_start = encoder.len();
+        encoder.encode_byte(PACKET_BIT_LONG | PACKET_BIT_FIXED_QUIC | pt.code() << 4);
+        encoder.encode_uint(4, QUIC_VERSION);
+        encoder.encode_vec(1, dcid);
+        encoder.encode_vec(1, scid);
+        Self {
+            encoder,
+            pn: u64::max_value(),
+            header: header_start..header_start,
+            offsets: PacketBuilderoffsets {
+                first_byte_mask: 0x0f,
+                pn: 0..0,
+                len: 0,
+            },
+        }
+    }
+
+    /// For an Initial packet, encode the token.
+    /// If you fail to do this, then you will not get a valid packet.
+    pub fn initial_token(&mut self, token: &[u8]) {
+        debug_assert_eq!(
+            self.encoder[self.header.start] & 0xb0,
+            PACKET_BIT_LONG | PACKET_TYPE_INITIAL << 4
+        );
+        self.encoder.encode_vvec(token);
+    }
+
+    /// Add a packet number of the given size.
+    /// For a long header packet, this also inserts a dummy length.
+    /// The length is filled in after calling `build`.
+    pub fn pn(&mut self, pn: PacketNumber, pn_len: usize) {
+        // Reserve space for a length in long headers.
+        if (self.encoder[self.header.start] & 0x80) == PACKET_BIT_LONG {
+            self.offsets.len = self.encoder.len();
+            self.encoder.encode(&[0; 2]);
+        }
+        // Encode the packet number and save its offset.
+        debug_assert!(pn_len <= 4 && pn_len > 0);
+        let pn_offset = self.encoder.len();
+        self.encoder.encode_uint(pn_len, pn);
+        self.offsets.pn = pn_offset..self.encoder.len();
+
+        // Now encode the packet number length and save the header length.
+        self.encoder[self.header.start] |= (pn_len - 1) as u8;
+        self.header.end = self.encoder.len();
+        self.pn = pn;
+    }
+
+    fn write_len(&mut self, expansion: usize) {
+        let len = self.encoder.len() - (self.offsets.len + 2) + expansion;
+        self.encoder[self.offsets.len] = 0x40 | ((len >> 8) & 0x3f) as u8;
+        self.encoder[self.offsets.len + 1] = (len & 0xff) as u8;
+    }
+
+    /// Build the packet and return the encoder.
+    pub fn build(mut self, crypto: &mut CryptoDxState) -> Res<Encoder> {
+        if self.offsets.len > 0 {
+            self.write_len(crypto.expansion());
+        }
+        let hdr = &self.encoder[self.header.clone()];
+        let body = &self.encoder[self.header.end..];
+        qdebug!(
+            "Packet build pn={} hdr={} body={}",
+            self.pn,
+            hex(hdr),
+            hex(body)
+        );
+        let ciphertext = crypto.encrypt(self.pn, hdr, body)?;
+
+        // Calculate the mask.
+        let offset = 4 - self.offsets.pn.len();
+        assert!(offset + SAMPLE_SIZE <= ciphertext.len());
+        let sample = &ciphertext[offset..offset + SAMPLE_SIZE];
+        let mask = crypto.compute_mask(sample)?;
+
+        // Apply the mask.
+        self.encoder[self.header.start] ^= mask[0] & self.offsets.first_byte_mask;
+        for (i, j) in (1..=self.offsets.pn.len()).zip(self.offsets.pn) {
+            self.encoder[j] ^= mask[i];
+        }
+
+        // Finally, cut off the plaintext and add back the ciphertext.
+        self.encoder.truncate(self.header.end);
+        self.encoder.encode(&ciphertext);
+        qdebug!("Packet built {}", hex(&self.encoder));
+        Ok(self.encoder)
+    }
+
+    /// Abort writing of this packet and return the encoder.
     #[must_use]
-    pub fn key_phase(&self) -> bool {
-        matches!(self, Self::Short(true))
+    pub fn abort(mut self) -> Encoder {
+        self.encoder.truncate(self.header.start);
+        self.encoder
     }
-}
 
-impl Default for PacketType {
-    fn default() -> Self {
-        Self::Short(false)
+    /// Work out if nothing was added after the header.
+    #[must_use]
+    pub fn is_empty(&self) -> bool {
+        self.encoder.len() == self.header.end
     }
-}
 
-pub type Version = u32;
-pub type PacketNumber = u64;
-
-#[derive(Default, Debug)]
-#[allow(clippy::module_name_repetitions)]
-pub struct PacketHdr {
-    pub tbyte: u8,
-    pub tipe: PacketType,
-    pub version: Option<Version>,
-    pub dcid: ConnectionId,
-    pub scid: Option<ConnectionId>,
-    pub pn: PacketNumber,
-    pub hdr_len: usize,
-    body_len: usize,
-}
-
-impl PacketHdr {
-    pub fn body_len(&self) -> usize {
-        self.body_len
+    /// Make a retry packet.
+    /// As this is a simple packet, this is just an associated function.
+    /// As Retry is odd (it has to be constructed with leading bytes),
+    /// this returns a Vec<u8> rather than building on an encoder.
+    pub fn retry(dcid: &[u8], scid: &[u8], token: &[u8], odcid: &[u8]) -> Res<Vec<u8>> {
+        let mut encoder = Encoder::default();
+        encoder.encode_vec(1, odcid);
+        let start = encoder.len();
+        encoder.encode_byte(
+            PACKET_BIT_LONG
+                | PACKET_BIT_FIXED_QUIC
+                | (PACKET_TYPE_RETRY << 4)
+                | (random(1)[0] & 0xf),
+        );
+        encoder.encode_uint(4, QUIC_VERSION);
+        encoder.encode_vec(1, dcid);
+        encoder.encode_vec(1, scid);
+        debug_assert_ne!(token.len(), 0);
+        encoder.encode(token);
+        let tag = RETRY_AEAD
+            .try_with(|aead| -> Res<Vec<u8>> {
+                let mut buf = vec![0; aead.borrow().expansion()];
+                Ok(aead.borrow().encrypt(0, &encoder, &[], &mut buf)?.to_vec())
+            })
+            .map_err(|e| {
+                qerror!("Unable to access Retry AEAD: {:?}", e);
+                Error::InternalError
+            })??;
+        encoder.encode(&tag);
+        let mut complete: Vec<u8> = encoder.into();
+        Ok(complete.split_off(start))
     }
-}
 
-// TODO(mt) test this.  It's a strict implementation of the spec,
-// but that doesn't mean we shouldn't test it.
-fn decode_pn(expected: PacketNumber, pn: u64, w: usize) -> PacketNumber {
-    let window = 1_u64 << (w * 8);
-    let candidate = (expected & !(window - 1)) | pn;
-    if candidate + (window / 2) <= expected {
-        candidate + window
-    } else if candidate > expected + (window / 2) {
-        match candidate.checked_sub(window) {
-            Some(pn_sub) => pn_sub,
-            None => candidate,
+    /// Make a Version Negotiation packet.
+    pub fn version_negotiation(dcid: &[u8], scid: &[u8]) -> Vec<u8> {
+        let mut encoder = Encoder::default();
+        let mut grease = random(5);
+        // This will not include the "QUIC bit" sometimes.  Intentionally.
+        encoder.encode_byte(PACKET_BIT_LONG | (grease[4] & 0x7f));
+        encoder.encode(&[0; 4]); // Zero version == VN.
+        encoder.encode_vec(1, dcid);
+        encoder.encode_vec(1, scid);
+        encoder.encode_uint(4, QUIC_VERSION);
+        // Add a greased version, using the randomness already generated.
+        for g in &mut grease[..4] {
+            *g = *g & 0xf0 | 0x0a;
         }
-    } else {
-        candidate
+        encoder.encode(&grease[0..4]);
+        encoder.into()
     }
 }
 
-fn decode_pnl(u: u8) -> usize {
-    assert!(u < 4); // This came from 2 bits
-    (u + 1) as usize
+impl Deref for PacketBuilder {
+    type Target = Encoder;
+
+    fn deref(&self) -> &Self::Target {
+        &self.encoder
+    }
 }
 
-/*
-  Short Header
+impl DerefMut for PacketBuilder {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.encoder
+    }
+}
 
-   0                   1                   2                   3
-   0 1 2 3 4 5 6 7 8 9 0 1 2 3 4 5 6 7 8 9 0 1 2 3 4 5 6 7 8 9 0 1
-  +-+-+-+-+-+-+-+-+
-  |0|1|S|R|R|K|P P|
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  |                Destination Connection ID (0..144)           ...
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  |                     Packet Number (8/16/24/32)              ...
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  |                     Protected Payload (*)                   ...
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
+impl Into<Encoder> for PacketBuilder {
+    fn into(self) -> Encoder {
+        self.encoder
+    }
+}
 
+/// PublicPacket holds information from packets that is public only.  This allows for
+/// processing of packets prior to decryption.
+pub struct PublicPacket<'a> {
+    /// The packet type.
+    packet_type: PacketType,
+    /// The recovered destination connection ID.
+    dcid: ConnectionIdRef<'a>,
+    /// The source connection ID, if this is a long header packet.
+    scid: Option<ConnectionIdRef<'a>>,
+    /// Any token that is included in the packet (Retry always has a token; Initial sometimes does).
+    /// This is empty when there is no token.
+    token: &'a [u8],
+    /// The size of the header, not including the packet number.
+    header_len: usize,
+    /// A reference to the entire packet, including the header.
+    data: &'a [u8],
+}
 
-  Long Header
+impl<'a> PublicPacket<'a> {
+    fn opt<T>(v: Option<T>) -> Res<T> {
+        if let Some(v) = v {
+            Ok(v)
+        } else {
+            Err(Error::NoMoreData)
+        }
+    }
 
-   0                   1                   2                   3
-   0 1 2 3 4 5 6 7 8 9 0 1 2 3 4 5 6 7 8 9 0 1 2 3 4 5 6 7 8 9 0 1
-  +-+-+-+-+-+-+-+-+
-  |1|1|T T|X X X X|
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  |                         Version (32)                          |
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  | DCID Len (8)  |
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  |               Destination Connection ID (0..160)            ...
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  | SCID Len (8)  |
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  |                 Source Connection ID (0..160)               ...
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-
-  Handshake
-  +-+-+-+-+-+-+-+-+
-  |1|1| 2 |R R|P P|
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  |                         Version (32)                          |
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  | DCID Len (8)  |
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  |               Destination Connection ID (0..160)            ...
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  | SCID Len (8)  |
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  |                 Source Connection ID (0..160)               ...
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  |                           Length (i)                        ...
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  |                    Packet Number (8/16/24/32)               ...
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  |                          Payload (*)                        ...
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-
-  Retry
-   0                   1                   2                   3
-   0 1 2 3 4 5 6 7 8 9 0 1 2 3 4 5 6 7 8 9 0 1 2 3 4 5 6 7 8 9 0 1
-  +-+-+-+-+-+-+-+-+
-  |1|1| 3 | Unused|
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  |                         Version (32)                          |
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  | DCID Len (8)  |
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  |               Destination Connection ID (0..160)            ...
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  | SCID Len (8)  |
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  |                 Source Connection ID (0..160)               ...
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  | ODCID Len (8) |
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  |          Original Destination Connection ID (0..160)        ...
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-  |                        Retry Token (*)                      ...
-  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-*/
-
-pub fn decode_packet_hdr(cid_parser: &dyn ConnectionIdDecoder, pd: &[u8]) -> Res<PacketHdr> {
-    macro_rules! d {
-        ($d:expr) => {
-            match $d {
-                Some(v) => v,
-                _ => return Err(Error::NoMoreData),
+    /// Decode the type-specific portions of a long header.
+    /// This includes reading the length and the remainder of the packet.
+    /// Returns a tuple of any token and the length of the header.
+    fn decode_type_specific(
+        decoder: &mut Decoder<'a>,
+        packet_type: PacketType,
+    ) -> Res<(&'a [u8], usize)> {
+        if packet_type == PacketType::Retry {
+            let header_len = decoder.offset();
+            let expansion = retry_expansion();
+            let token = Self::opt(decoder.decode(decoder.remaining() - expansion))?;
+            if token.is_empty() {
+                return Err(Error::InvalidPacket);
             }
+            Self::opt(decoder.decode(expansion))?;
+            return Ok((token, header_len));
+        }
+        let token = if packet_type == PacketType::Initial {
+            Self::opt(decoder.decode_vvec())?
+        } else {
+            &[]
         };
+        let len = Self::opt(decoder.decode_varint())?;
+        let header_len = decoder.offset();
+        let _body = Self::opt(decoder.decode(usize::try_from(len)?))?;
+        Ok((token, header_len))
     }
 
-    let mut p = PacketHdr::default();
-    let mut d = Decoder::from(pd);
+    /// Decode the common parts of a packet.  This provides minimal parsing and validation.
+    /// Returns a tuple of a `PublicPacket` and a slice with any remainder from the datagram.
+    pub fn decode(data: &'a [u8], dcid_decoder: &dyn ConnectionIdDecoder) -> Res<(Self, &'a [u8])> {
+        let mut decoder = Decoder::new(data);
+        let first = Self::opt(decoder.decode_byte())?;
 
-    // Get the type byte
-    p.tbyte = d!(d.decode_byte());
-    if (p.tbyte & 0x80) == 0 {
-        if p.tbyte & 0x40 == 0 {
+        if first & 0x80 == PACKET_BIT_SHORT {
+            return if first & 0x40 == PACKET_BIT_FIXED_QUIC {
+                let dcid = Self::opt(dcid_decoder.decode_cid(&mut decoder))?;
+                let header_len = decoder.offset();
+                Ok((
+                    Self {
+                        packet_type: PacketType::Short,
+                        dcid,
+                        scid: None,
+                        token: &[],
+                        header_len,
+                        data,
+                    },
+                    &[],
+                ))
+            } else {
+                Err(Error::InvalidPacket)
+            };
+        }
+
+        // Generic long header.
+        let version = Version::try_from(Self::opt(decoder.decode_uint(4))?).unwrap();
+        let dcid = ConnectionIdRef::from(Self::opt(decoder.decode_vec(1))?);
+        let scid = ConnectionIdRef::from(Self::opt(decoder.decode_vec(1))?);
+
+        // Version negotiation.
+        if version == 0 {
+            return Ok((
+                Self {
+                    packet_type: PacketType::VersionNegotiation,
+                    dcid,
+                    scid: Some(scid),
+                    token: &[],
+                    header_len: decoder.offset(),
+                    data,
+                },
+                &[],
+            ));
+        }
+
+        // Check that this is a long header from this version.
+        if version != QUIC_VERSION {
+            return Ok((
+                Self {
+                    packet_type: PacketType::OtherVersion,
+                    dcid,
+                    scid: Some(scid),
+                    token: &[],
+                    header_len: decoder.offset(),
+                    data,
+                },
+                &[],
+            ));
+        }
+        if (first & PACKET_BIT_FIXED_QUIC) != PACKET_BIT_FIXED_QUIC {
             return Err(Error::InvalidPacket);
         }
-
-        // Short Header.
-        p.tipe = PacketType::Short((p.tbyte & 4) != 0);
-        let cid = d!(cid_parser.decode_cid(&mut d));
-        p.dcid = ConnectionId::from(&cid); // TODO(mt) unnecessary copy
-        p.hdr_len = pd.len() - d.remaining();
-        p.body_len = d.remaining();
-        return Ok(p);
-    }
-
-    let version = d!(d.decode_uint(4)) as u32;
-    p.version = Some(version);
-    p.dcid = ConnectionId::from(d!(d.decode_vec(1)));
-    p.scid = Some(ConnectionId::from(d!(d.decode_vec(1))));
-
-    if version == 0 {
-        let mut vns = vec![];
-        while d.remaining() > 0 {
-            vns.push(d!(d.decode_uint(4)) as u32);
-        }
-        p.tipe = PacketType::VN(vns);
-        // No need to set hdr_length and body_length
-        // because we won't need them.
-        return Ok(p);
-    } else {
-        if p.tbyte & 0x40 == 0 {
-            return Err(Error::InvalidPacket);
-        }
-
-        p.tipe = match (p.tbyte >> 4) & 0x3 {
-            // TODO(ekr@rtfm.com): Check the 0 bits.
-            PACKET_TYPE_INITIAL => {
-                PacketType::Initial(d!(d.decode_vvec()).to_vec()) // TODO(mt) unnecessary copy
-            }
-            PACKET_TYPE_0RTT => PacketType::ZeroRTT,
+        let packet_type = match (first >> 4) & 3 {
+            PACKET_TYPE_INITIAL => PacketType::Initial,
+            PACKET_TYPE_0RTT => PacketType::ZeroRtt,
             PACKET_TYPE_HANDSHAKE => PacketType::Handshake,
-            PACKET_TYPE_RETRY => {
-                // let odcid = ConnectionId(d!(d.decode_vec(1)).to_vec()); // TODO(mt) unnecessary copy
-                let odcid = ConnectionId::from(&[][..]);
-                let mut token = d.decode_remainder().to_vec(); // TODO(mt) unnecessary copy
-                token.truncate(token.len() - 16); // TODO(mt) check the tag
-                p.tipe = PacketType::Retry { odcid, token };
-                return Ok(p);
-            }
+            PACKET_TYPE_RETRY => PacketType::Retry,
             _ => unreachable!(),
         };
+
+        // The type-specific code includes a token.  This consumes the remainder of the packet.
+        let (token, header_len) = Self::decode_type_specific(&mut decoder, packet_type)?;
+        let end = data.len() - decoder.remaining();
+        let (data, remainder) = data.split_at(end);
+        Ok((
+            Self {
+                packet_type,
+                dcid,
+                scid: Some(scid),
+                token,
+                header_len,
+                data,
+            },
+            remainder,
+        ))
     }
 
-    p.body_len = usize::try_from(d!(d.decode_varint()))?;
-    if p.body_len > d.remaining() {
-        return Err(Error::InvalidPacket);
+    /// Validate the given packet as though it were a retry.
+    pub fn is_valid_retry(&self, odcid: &ConnectionId) -> bool {
+        if self.packet_type != PacketType::Retry {
+            return false;
+        }
+        let expansion = retry_expansion();
+        if self.data.len() <= expansion {
+            return false;
+        }
+        let (header, tag) = self.data.split_at(self.data.len() - expansion);
+        let mut encoder = Encoder::with_capacity(self.data.len());
+        encoder.encode_vec(1, odcid);
+        encoder.encode(header);
+        RETRY_AEAD
+            .try_with(|aead| -> bool {
+                let mut buf = vec![0; expansion];
+                if let Ok(v) = aead.borrow().decrypt(0, &encoder, tag, &mut buf) {
+                    v.is_empty()
+                } else {
+                    false
+                }
+            })
+            .unwrap_or_else(|e| {
+                qerror!("Unable to access Retry AEAD: {:?}", e);
+                false
+            })
     }
-    p.hdr_len = pd.len() - d.remaining();
 
-    Ok(p)
-}
+    pub fn is_valid_initial(&self) -> bool {
+        // Packet has to be an initial, with a DCID of 8 bytes, or a token.
+        // Assume that the Server class validates the token.
+        self.packet_type == PacketType::Initial
+            && (self.dcid().len() >= 8 || !self.token.is_empty())
+    }
 
-pub fn decrypt_packet_hdr<'p>(
-    crypto: &mut CryptoDxState,
-    hdr: &mut PacketHdr,
-    pkt: &'p [u8],
-) -> Res<(Vec<u8>, &'p [u8])> {
-    assert!(!matches!(
-        hdr.tipe,
-        PacketType::Retry{..} | PacketType::VN(_)
-    ));
+    pub fn packet_type(&self) -> PacketType {
+        self.packet_type
+    }
 
-    qtrace!("unmask hdr={}", hex(&pkt[..hdr.hdr_len + 4]));
+    pub fn dcid(&self) -> &ConnectionIdRef<'a> {
+        &self.dcid
+    }
 
-    let sample_offset = hdr.hdr_len + 4;
-    let mask = if let Some(sample) = pkt.get(sample_offset..(sample_offset + SAMPLE_SIZE)) {
-        crypto.compute_mask(sample)
-    } else {
-        Err(Error::NoMoreData)
-    }?;
+    pub fn scid(&self) -> &ConnectionIdRef<'a> {
+        self.scid
+            .as_ref()
+            .expect("should only be called for long header packets")
+    }
 
-    // Un-mask the leading byte.
-    debug_assert_eq!(hdr.tbyte, pkt[0]);
-    hdr.tbyte ^= mask[0]
-        & match hdr.tipe {
-            PacketType::Short(key_phase) => {
-                let flip = (mask[0] & 4) != 0;
-                hdr.tipe = PacketType::Short(key_phase ^ flip);
-                0x1f
+    pub fn token(&self) -> &'a [u8] {
+        self.token
+    }
+
+    fn decode_pn(expected: PacketNumber, pn: u64, w: usize) -> PacketNumber {
+        let window = 1_u64 << (w * 8);
+        let candidate = (expected & !(window - 1)) | pn;
+        if candidate + (window / 2) <= expected {
+            candidate + window
+        } else if candidate > expected + (window / 2) {
+            match candidate.checked_sub(window) {
+                Some(pn_sub) => pn_sub,
+                None => candidate,
             }
-            _ => 0x0f,
-        };
-    let pn_len = decode_pnl(hdr.tbyte & 0x3);
-
-    // Make a copy of the header to work on.
-    let mut hdrbytes = pkt[0..(hdr.hdr_len + pn_len)].to_vec();
-    hdrbytes[0] = hdr.tbyte;
-
-    // Unmask the PN.
-    let mut pn_encoded: u64 = 0;
-    for i in 0..pn_len {
-        hdrbytes[hdr.hdr_len + i] ^= mask[1 + i];
-        pn_encoded <<= 8;
-        pn_encoded += u64::from(hdrbytes[hdr.hdr_len + i]);
+        } else {
+            candidate
+        }
     }
 
-    qtrace!("unmasked hdr={}", hex(&hdrbytes));
-    hdr.hdr_len += pn_len;
-    hdr.body_len -= pn_len;
+    /// Decrypt the header of the packet.
+    fn decrypt_header(
+        &self,
+        crypto: &mut CryptoDxState,
+    ) -> Res<(bool, PacketNumber, Vec<u8>, &'a [u8])> {
+        assert_ne!(self.packet_type, PacketType::Retry);
+        assert_ne!(self.packet_type, PacketType::VersionNegotiation);
 
-    hdr.pn = decode_pn(crypto.next_pn(), pn_encoded, pn_len);
-    Ok((hdrbytes, &pkt[hdr.hdr_len..hdr.hdr_len + hdr.body_len]))
+        qtrace!("unmask hdr={}", hex(&self.data[..self.header_len + 4]));
+
+        let sample_offset = self.header_len + 4;
+        let mask = if let Some(sample) = self.data.get(sample_offset..(sample_offset + SAMPLE_SIZE))
+        {
+            crypto.compute_mask(sample)
+        } else {
+            Err(Error::NoMoreData)
+        }?;
+
+        // Un-mask the leading byte.
+        let first = self.data[0]
+            ^ (mask[0]
+                & if self.packet_type == PacketType::Short {
+                    0x1f
+                } else {
+                    0x0f
+                });
+        let pn_len = usize::from((first & 0x3) + 1);
+
+        // Make a copy of the header to work on.
+        let mut hdrbytes = self.data[..self.header_len + pn_len].to_vec();
+        hdrbytes[0] = first;
+
+        // Unmask the PN.
+        let mut pn_encoded: u64 = 0;
+        for i in 0..pn_len {
+            hdrbytes[self.header_len + i] ^= mask[1 + i];
+            pn_encoded <<= 8;
+            pn_encoded += u64::from(hdrbytes[self.header_len + i]);
+        }
+
+        qtrace!("unmasked hdr={}", hex(&hdrbytes));
+
+        let key_phase = self.packet_type == PacketType::Short
+            && (first & PACKET_BIT_KEY_PHASE) == PACKET_BIT_KEY_PHASE;
+        let pn = Self::decode_pn(crypto.next_pn(), pn_encoded, pn_len);
+        Ok((
+            key_phase,
+            pn,
+            hdrbytes,
+            &self.data[self.header_len + pn_len..],
+        ))
+    }
+
+    pub fn decrypt(&self, crypto: &mut CryptoStates, release_at: Instant) -> Res<DecryptedPacket> {
+        let space = PNSpace::from(self.packet_type);
+        // This has to work in two stages because we need to remove header protection
+        // before picking the keys to use.
+        if let Some(rx) = crypto.rx_hp(space) {
+            // Note that this will dump early, which creates a side-channel.
+            // This is OK in this case because we the only reason this can
+            // fail is if the cryptographic module is bad or the packet is
+            // too small (which is public information).
+            let (key_phase, pn, header, body) = self.decrypt_header(rx)?;
+            qtrace!([rx], "decoded header: {:?}", header);
+            if let Some(rx) = crypto.rx(space, key_phase) {
+                let d = rx.decrypt(pn, &header, body)?;
+                // If this is the first packet ever successfully decrypted
+                // using `rx`, make sure to initiate a key update.
+                if rx.needs_update() {
+                    crypto.key_update_received(release_at)?;
+                }
+                crypto.check_pn_overlap()?;
+                Ok(DecryptedPacket {
+                    pt: self.packet_type,
+                    pn,
+                    data: d,
+                })
+            } else {
+                Err(Error::DecryptError)
+            }
+        } else {
+            Err(Error::DecryptError)
+        }
+    }
 }
 
-pub fn decrypt_packet_body(
-    crypto: &mut CryptoDxState,
+impl fmt::Debug for PublicPacket<'_> {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(
+            f,
+            "{:?}: {} {}",
+            self.packet_type(),
+            hex(&self.data[..self.header_len]),
+            hex(&self.data[self.header_len..])
+        )
+    }
+}
+
+pub struct DecryptedPacket {
+    pt: PacketType,
     pn: PacketNumber,
-    hdrbytes: &[u8],
-    body: &[u8],
-) -> Res<Vec<u8>> {
-    Ok(crypto.decrypt(pn, hdrbytes, body)?)
+    data: Vec<u8>,
 }
 
-#[cfg(disabled)]
+impl DecryptedPacket {
+    pub fn packet_type(&self) -> PacketType {
+        self.pt
+    }
+
+    pub fn pn(&self) -> PacketNumber {
+        self.pn
+    }
+}
+
+impl Deref for DecryptedPacket {
+    type Target = [u8];
+
+    fn deref(&self) -> &Self::Target {
+        &self.data[..]
+    }
+}
+
+#[cfg(test)]
 mod tests {
     use super::*;
-    use neqo_common::matches;
+    use crate::crypto::{CryptoDxDirection, CryptoDxState};
+    use crate::FixedConnectionIdManager;
+    use neqo_common::Encoder;
     use test_fixture::fixture_init;
 
-    const TEST_BODY: [u8; 6] = [0x01, 0x23, 0x45, 0x67, 0x89, 0x10];
+    const CLIENT_CID: &[u8] = &[0x83, 0x94, 0xc8, 0xf0, 0x3e, 0x51, 0x57, 0x08];
+    const SERVER_CID: &[u8] = &[0xf0, 0x67, 0xa5, 0x50, 0x2a, 0x42, 0x62, 0xb5];
 
-    struct TestFixture {}
-
-    const AEAD_MASK: u8 = 0;
-    const AUTH_TAG_LEN: usize = 16;
-
-    impl TestFixture {
-        fn auth_tag(hdr: &[u8], body: &[u8]) -> [u8; AUTH_TAG_LEN] {
-            [0xa; AUTH_TAG_LEN]
-        }
-    }
-
-    impl HeaderProtectionMask for TestFixture {
-        fn compute_mask(&self, sample: &[u8]) -> Res<Vec<u8>> {
-            Ok(vec![0xa5; 5])
-        }
-
-        #[must_use]
-        fn next_pn(&self) -> PacketNumber {
-            0
-        }
-    }
-
-    impl Unprotector for TestFixture {
-        fn decrypt(&mut self, pn: PacketNumber, hdr: &[u8], body: &[u8]) -> Res<Vec<u8>> {
-            let mut pt = body.to_vec();
-
-            for i in &mut pt {
-                *i ^= AEAD_MASK;
-            }
-            let pt_len = pt.len() - AUTH_TAG_LEN;
-            let at = Self::auth_tag(hdr, &pt[0..pt_len]);
-            for i in 0..16 {
-                if at[i] != pt[pt_len + i] {
-                    return Err(Error::DecryptError);
-                }
-            }
-            Ok(pt[0..pt_len].to_vec())
-        }
-    }
-
-    impl Protector for TestFixture {
-        fn encrypt(&mut self, pn: PacketNumber, hdr: &[u8], body: &[u8]) -> Res<Vec<u8>> {
-            let tag = Self::auth_tag(hdr, body);
-            let mut enc = Encoder::with_capacity(body.len() + tag.len());
-            enc.encode(body);
-            enc.encode(&tag);
-            for i in 0..enc.len() {
-                enc[i] ^= AEAD_MASK;
-            }
-
-            Ok(enc.into())
-        }
-        fn expansion(&self) -> usize {
-            AUTH_TAG_LEN
-        }
-    }
-
-    impl ConnectionIdDecoder for TestFixture {
-        fn decode_cid(&self, dec: &mut Decoder) -> Option<ConnectionId> {
-            dec.decode(5).map(ConnectionId::from)
-        }
-    }
-
-    fn default_hdr() -> PacketHdr {
-        PacketHdr {
-            tbyte: 0,
-            tipe: PacketType::Short(false),
-            version: Some(31),
-            dcid: ConnectionId(vec![1, 2, 3, 4, 5]),
-            scid: None,
-            pn: 0x0505,
-            hdr_len: 0,
-            body_len: 0,
-        }
-    }
-
-    fn assert_headers_equal(left: &PacketHdr, right: &PacketHdr) {
-        assert_eq!(left.tipe, right.tipe);
-        assert_eq!(left.dcid, right.dcid);
-        assert_eq!(left.scid, right.scid);
-        assert_eq!(left.pn, right.pn);
-    }
-
-    fn test_decrypt_packet(f: &mut TestFixture, packet: Vec<u8>) -> Res<(PacketHdr, Vec<u8>)> {
-        let mut phdr = decode_packet_hdr(f, &packet)?;
-        let (hdr, body) = decrypt_packet_hdr(f, &mut phdr, &packet)?;
-        let payload = decrypt_packet_body(f, phdr.pn, &hdr, body)?;
-        Ok((phdr, payload))
-    }
-
-    fn test_encrypt_decrypt(f: &mut TestFixture, hdr: &mut PacketHdr, body: &[u8]) -> PacketHdr {
-        let packet = encode_packet(f, hdr, &TEST_BODY);
-        let (dec_hdr, dec_body) = test_decrypt_packet(f, packet).unwrap();
-        assert_headers_equal(&hdr, &dec_hdr);
-        assert_eq!(body.to_vec(), dec_body);
-        dec_hdr
-    }
-
-    #[test]
-    fn test_short_packet() {
-        let mut f = TestFixture {};
-        let mut hdr = default_hdr();
-        test_encrypt_decrypt(&mut f, &mut hdr, &TEST_BODY);
-    }
-
-    #[test]
-    fn test_short_packet_damaged() {
-        let mut f = TestFixture {};
-        let hdr = default_hdr();
-        let mut packet = encode_packet(&mut f, &hdr, &TEST_BODY);
-        let plen = packet.len();
-        packet[plen - 1] ^= 0x7;
-        assert!(test_decrypt_packet(&mut f, packet).is_err());
-    }
-
-    #[test]
-    fn test_handshake_packet() {
-        let mut f = TestFixture {};
-        let mut hdr = default_hdr();
-        hdr.tipe = PacketType::Handshake;
-        hdr.scid = Some(ConnectionId(vec![9, 8, 7, 6, 5, 4, 3, 2]));
-        test_encrypt_decrypt(&mut f, &mut hdr, &TEST_BODY);
-    }
-
-    #[test]
-    fn test_handshake_packet_damaged() {
-        let mut f = TestFixture {};
-        let mut hdr = default_hdr();
-        hdr.tipe = PacketType::Handshake;
-        hdr.scid = Some(ConnectionId(vec![9, 8, 7, 6, 5, 4, 3, 2]));
-        let mut packet = encode_packet(&mut f, &hdr, &TEST_BODY);
-        let plen = packet.len();
-        packet[plen - 1] ^= 0x7;
-        assert!(test_decrypt_packet(&mut f, packet).is_err());
-    }
-
-    #[test]
-    fn test_initial_packet() {
-        let mut f = TestFixture {};
-        let mut hdr = default_hdr();
-        let tipe = PacketType::Initial(vec![0x0, 0x0, 0x0, 0x0]);
-        hdr.tipe = tipe;
-        hdr.scid = Some(ConnectionId(vec![9, 8, 7, 6, 5, 4, 3, 2]));
-        test_encrypt_decrypt(&mut f, &mut hdr, &TEST_BODY);
-    }
-
-    #[test]
-    fn test_initial_packet_damaged() {
-        let mut f = TestFixture {};
-        let mut hdr = default_hdr();
-        hdr.tipe = PacketType::Initial(vec![0x0, 0x0, 0x0, 0x0]);
-        hdr.scid = Some(ConnectionId(vec![9, 8, 7, 6, 5, 4, 3, 2]));
-        let mut packet = encode_packet(&mut f, &hdr, &TEST_BODY);
-        let plen = packet.len();
-        packet[plen - 1] ^= 0x7;
-        assert!(test_decrypt_packet(&mut f, packet).is_err());
-    }
-
-    #[test]
-    fn test_retry() {
+    /// In most of these tests, anything will do.  This is that "anything".
+    fn default_protector() -> CryptoDxState {
         fixture_init();
-        let mut hdr = default_hdr();
-        hdr.tipe = PacketType::Retry {
-            odcid: ConnectionId(vec![9, 8, 7, 6, 5, 4, 3, 2]),
-            token: vec![99, 88, 77, 66, 55, 44, 33],
-        };
-        hdr.scid = Some(ConnectionId(vec![1, 2, 3, 4, 5]));
-        let packet = encode_retry(&hdr);
-        let f = TestFixture {};
-        let decoded = decode_packet_hdr(&f, &packet).expect("should decode");
-        assert_eq!(decoded.tipe, hdr.tipe);
-        assert_eq!(decoded.version, hdr.version);
-        assert_eq!(decoded.dcid, hdr.dcid);
-        assert_eq!(decoded.scid, hdr.scid);
+        CryptoDxState::new_initial(CryptoDxDirection::Write, "server in", CLIENT_CID)
     }
 
     #[test]
-    fn generate_initial_cid() {
+    fn sample_server_initial() {
+        const SAMPLE_PAYLOAD: &[u8] = &[
+            0x0d, 0x00, 0x00, 0x00, 0x00, 0x18, 0x41, 0x0a, 0x02, 0x00, 0x00, 0x56, 0x03, 0x03,
+            0xee, 0xfc, 0xe7, 0xf7, 0xb3, 0x7b, 0xa1, 0xd1, 0x63, 0x2e, 0x96, 0x67, 0x78, 0x25,
+            0xdd, 0xf7, 0x39, 0x88, 0xcf, 0xc7, 0x98, 0x25, 0xdf, 0x56, 0x6d, 0xc5, 0x43, 0x0b,
+            0x9a, 0x04, 0x5a, 0x12, 0x00, 0x13, 0x01, 0x00, 0x00, 0x2e, 0x00, 0x33, 0x00, 0x24,
+            0x00, 0x1d, 0x00, 0x20, 0x9d, 0x3c, 0x94, 0x0d, 0x89, 0x69, 0x0b, 0x84, 0xd0, 0x8a,
+            0x60, 0x99, 0x3c, 0x14, 0x4e, 0xca, 0x68, 0x4d, 0x10, 0x81, 0x28, 0x7c, 0x83, 0x4d,
+            0x53, 0x11, 0xbc, 0xf3, 0x2b, 0xb9, 0xda, 0x1a, 0x00, 0x2b, 0x00, 0x02, 0x03, 0x04,
+        ];
+        const EXPECTED: &[u8] = &[
+            0xc9, 0xff, 0x00, 0x00, 0x18, 0x00, 0x08, 0xf0, 0x67, 0xa5, 0x50, 0x2a, 0x42, 0x62,
+            0xb5, 0x00, 0x40, 0x74, 0x16, 0x8b, 0xf2, 0x2b, 0x70, 0x02, 0x59, 0x6f, 0x99, 0xae,
+            0x67, 0xab, 0xf6, 0x5a, 0x58, 0x52, 0xf5, 0x4f, 0x58, 0xc3, 0x7c, 0x80, 0x86, 0x82,
+            0xe2, 0xe4, 0x04, 0x92, 0xd8, 0xa3, 0x89, 0x9f, 0xb0, 0x4f, 0xc0, 0xaf, 0xe9, 0xaa,
+            0xbc, 0x87, 0x67, 0xb1, 0x8a, 0x0a, 0xa4, 0x93, 0x53, 0x74, 0x26, 0x37, 0x3b, 0x48,
+            0xd5, 0x02, 0x21, 0x4d, 0xd8, 0x56, 0xd6, 0x3b, 0x78, 0xce, 0xe3, 0x7b, 0xc6, 0x64,
+            0xb3, 0xfe, 0x86, 0xd4, 0x87, 0xac, 0x7a, 0x77, 0xc5, 0x30, 0x38, 0xa3, 0xcd, 0x32,
+            0xf0, 0xb5, 0x00, 0x4d, 0x9f, 0x57, 0x54, 0xc4, 0xf7, 0xf2, 0xd1, 0xf3, 0x5c, 0xf3,
+            // -- For draft-25, change the 0x18 above and use these lines:
+            // 0xf7, 0x11, 0x63, 0x51, 0xc9, 0x2b, 0x99, 0xc8, 0xae, 0x58, 0x33, 0x22, 0x5c, 0xb5,
+            // 0x18, 0x55, 0x20, 0xd6, 0x1e, 0x68, 0xcf, 0x5f,
+            // -- These are draft-24 values:
+            0xf7, 0x11, 0x63, 0x51, 0xc9, 0x2b, 0x58, 0x4d, 0x2d, 0xdd, 0x6c, 0x26, 0xc7, 0x8a,
+            0x84, 0x07, 0xf0, 0x9e, 0xfd, 0xa4, 0xa3, 0x08,
+        ];
+
+        let mut prot = default_protector();
+
+        // The spec uses PN=1, but our crypto refuses to skip packet numbers.
+        // So burn an encryption:
+        let burn = prot.encrypt(0, &[], &[]).expect("burn OK");
+        assert_eq!(burn.len(), prot.expansion());
+
+        let mut builder = PacketBuilder::long(
+            Encoder::new(),
+            PacketType::Initial,
+            &ConnectionId::from(&[][..]),
+            &ConnectionId::from(SERVER_CID),
+        );
+        builder.initial_token(&[]);
+        builder.pn(1, 2);
+        builder.encode(&SAMPLE_PAYLOAD);
+        let packet = builder.build(&mut prot).expect("build");
+        assert_eq!(&packet[..], EXPECTED);
+    }
+
+    #[test]
+    fn build_short() {
+        let mut builder =
+            PacketBuilder::short(Encoder::new(), true, &ConnectionId::from(SERVER_CID));
+        builder.pn(0, 1);
+        builder.encode(&[0; 3]); // Enough payload for sampling.
+        let packet = builder.build(&mut default_protector()).expect("build");
+        assert_eq!(packet.len(), 29);
+    }
+
+    #[test]
+    fn build_two() {
+        let mut prot = default_protector();
+        let mut builder = PacketBuilder::long(
+            Encoder::new(),
+            PacketType::Handshake,
+            &ConnectionId::from(SERVER_CID),
+            &ConnectionId::from(CLIENT_CID),
+        );
+        builder.pn(0, 1);
+        builder.encode(&[0; 3]);
+        let encoder = builder.build(&mut prot).expect("build");
+        assert_eq!(encoder.len(), 45);
+        let first = encoder.clone();
+
+        let mut builder = PacketBuilder::short(encoder, false, &ConnectionId::from(SERVER_CID));
+        builder.pn(1, 3);
+        builder.encode(&[0]); // Minimal size (packet number is big enough).
+        let encoder = builder.build(&mut prot).expect("build");
+        assert_eq!(
+            &first[..],
+            &encoder[..first.len()],
+            "the first packet should be a prefix"
+        );
+        assert_eq!(encoder.len(), 45 + 29);
+    }
+
+    #[test]
+    fn build_abort() {
+        let mut builder = PacketBuilder::long(
+            Encoder::new(),
+            PacketType::Initial,
+            &ConnectionId::from(&[][..]),
+            &ConnectionId::from(SERVER_CID),
+        );
+        builder.initial_token(&[]);
+        builder.pn(1, 2);
+        let encoder = builder.abort();
+        assert!(encoder.is_empty());
+    }
+
+    const SAMPLE_RETRY: &[u8] = &[
+        0xff, 0xff, 0x00, 0x00, 0x18, 0x00, 0x08, 0xf0, 0x67, 0xa5, 0x50, 0x2a, 0x42, 0x62, 0xb5,
+        0x74, 0x6f, 0x6b, 0x65, 0x6e, 0x43, 0xe0, 0x42, 0xc5, 0xcc, 0x3c, 0x5d, 0xa7, 0x31, 0xee,
+        0xc9, 0xa9, 0xbc, 0x3c, 0xab,
+        0x32,
+        // Draft-25 values:
+        // 0xff, 0xff, 0x00, 0x00, 0x19, 0x00, 0x08, 0xf0, 0x67, 0xa5, 0x50, 0x2a, 0x42, 0x62, 0xb5, 0x74, 0x6f, 0x6b, 0x65, 0x6e, 0x1e, 0x5e, 0xc5, 0xb0, 0x14, 0xcb, 0xb1, 0xf0, 0xfd, 0x93, 0xdf, 0x40, 0x48, 0xc4, 0x46, 0xa6,
+    ];
+
+    #[test]
+    fn build_retry() {
         fixture_init();
-        for i in 0..100 {
-            let cid = ConnectionId::generate_initial();
-            if !matches!(cid.len(), 8..=20) {
-                panic!("connection ID {:?}", cid);
-            }
+        let retry = PacketBuilder::retry(&[], SERVER_CID, b"token", CLIENT_CID).unwrap();
+
+        let (packet, remainder) =
+            PublicPacket::decode(&retry, &FixedConnectionIdManager::new(5)).unwrap();
+        assert!(packet.is_valid_retry(&ConnectionId::from(CLIENT_CID)));
+        assert!(remainder.is_empty());
+
+        // The builder adds randomness, which makes expectations hard.
+        // So only do a full check when that randomness matches up.
+        if retry[0] == SAMPLE_RETRY[0] {
+            assert_eq!(&retry, &SAMPLE_RETRY);
+        } else {
+            // Otherwise, just check that the header is OK.
+            assert_eq!(retry[0] & 0xf0, 0xf0);
+            let header_range = 1..retry.len() - 16;
+            assert_eq!(&retry[header_range.clone()], &SAMPLE_RETRY[header_range]);
         }
+    }
+
+    #[test]
+    fn build_retry_multiple() {
+        // Run the build_retry test a few times.
+        // This increases the chance that the full comparison happens.
+        for _ in 0..32 {
+            build_retry();
+        }
+    }
+
+    /// Check some packets that are clearly not valid Retry packets.
+    #[test]
+    fn invalid_retry() {
+        fixture_init();
+        let cid_mgr = FixedConnectionIdManager::new(5);
+        let odcid = ConnectionId::from(CLIENT_CID);
+
+        assert!(PublicPacket::decode(&[], &cid_mgr).is_err());
+
+        let (packet, remainder) = PublicPacket::decode(SAMPLE_RETRY, &cid_mgr).unwrap();
+        assert!(remainder.is_empty());
+        assert!(packet.is_valid_retry(&odcid));
+
+        let mut damaged_retry = SAMPLE_RETRY.to_vec();
+        let last = damaged_retry.len() - 1;
+        damaged_retry[last] ^= 66;
+        let (packet, remainder) = PublicPacket::decode(&damaged_retry, &cid_mgr).unwrap();
+        assert!(remainder.is_empty());
+        assert!(!packet.is_valid_retry(&odcid));
+
+        damaged_retry.truncate(last);
+        let (packet, remainder) = PublicPacket::decode(&damaged_retry, &cid_mgr).unwrap();
+        assert!(remainder.is_empty());
+        assert!(!packet.is_valid_retry(&odcid));
+
+        // An invalid token should be rejected sooner.
+        damaged_retry.truncate(last - 4);
+        assert!(PublicPacket::decode(&damaged_retry, &cid_mgr).is_err());
+
+        damaged_retry.truncate(last - 1);
+        assert!(PublicPacket::decode(&damaged_retry, &cid_mgr).is_err());
+    }
+
+    const SAMPLE_VN: &[u8] = &[
+        0x80, 0x00, 0x00, 0x00, 0x00, 0x08, 0xf0, 0x67, 0xa5, 0x50, 0x2a, 0x42, 0x62, 0xb5, 0x08,
+        0x83, 0x94, 0xc8, 0xf0, 0x3e, 0x51, 0x57, 0x08, 0xff, 0x00, 0x00, 0x18, 0x0a, 0x0a, 0x0a,
+        0x0a,
+    ];
+
+    #[test]
+    fn build_vn() {
+        fixture_init();
+        let mut vn = PacketBuilder::version_negotiation(SERVER_CID, CLIENT_CID);
+        // Erase randomness from greasing...
+        assert_eq!(vn.len(), SAMPLE_VN.len());
+        vn[0] &= 0x80;
+        for v in vn.iter_mut().skip(SAMPLE_VN.len() - 4) {
+            *v &= 0x0f;
+        }
+        assert_eq!(&vn, &SAMPLE_VN);
+    }
+
+    #[test]
+    fn parse_vn() {
+        let (packet, remainder) =
+            PublicPacket::decode(SAMPLE_VN, &FixedConnectionIdManager::new(5)).unwrap();
+        assert!(remainder.is_empty());
+        assert_eq!(&packet.dcid[..], SERVER_CID);
+        assert!(packet.scid.is_some());
+        assert_eq!(&packet.scid.unwrap()[..], CLIENT_CID);
     }
 }
